+++ lib/src/focus.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:xly/src/button.dart';
 
@@ -148,7 +147,7 @@ class _XlyFocusableWidgetState extends State<_XlyFocusableWidget> {
               color: _hasFocus || _isHovering
                   ? widget.focusedBorderColor
                   : Colors.transparent,
-              width: widget.borderWidth.w,
+              width: widget.borderWidth,
             ),
           ),
           child: widget.child,